test: isolate DCP stripe geometry coverage
